requirements.txt
praw==7.8.1
supabase==2.9.1
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.24.3
requests==2.31.0
regex==2023.10.3
langdetect==1.0.9
fasttext==0.9.2
indic-nlp-library==0.81
mahanlp==0.9
scikit-learn==1.3.2
polyglot==16.7.4